# Remove polymer embeds