--- a/web/embed-dart.html
+++ b/web/embed-dart.html
@@ -1,69 +1,194 @@
 <!DOCTYPE html>
 
-<!-- Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
-     for details. All rights reserved. Use of this source code is governed by a
-     BSD-style license that can be found in the LICENSE file. This version is
-     not finalized. -->
+<!-- Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
+for details. All rights reserved. Use of this source code is governed by a
+BSD-style license that can be found in the LICENSE file. -->
 
 <html>
-  <head>
+<head>
     <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"/>
-    <meta name="apple-mobile-web-app-capable" content="yes">
+    <link rel="shortcut icon" href="/dart-64.png">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
     <meta name="mobile-web-app-capable" content="yes">
-    <meta name="scaffolded-by" content="https://github.com/google/stagehand">
+    <meta name="apple-mobile-web-app-capable" content="yes">
+
     <title>DartPad</title>
 
-    <script src="packages/dart_pad/bower/webcomponentsjs/webcomponents-lite.min.js"></script>
-
-    <link rel="import" href="scripts/embed_components.html"/>
-
-    <link href='text-fonts/inconsolata.css'
+    <link href='https://fonts.googleapis.com/css?family=Roboto'
           rel='stylesheet' type='text/css'>
-    <!--<link href='https://fonts.googleapis.com/css?family=Inconsolata:400,700'
+    <link href='https://fonts.googleapis.com/css?family=Inconsolata:400,700'
           rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto">-->
+
+    <link rel="stylesheet" type="text/css" href="../packages/octicons_css/octicons_4.4.0.css">
 
     <!-- codemirror -->
-    <link href="packages/codemirror/css/codemirror.css" rel="stylesheet" media="screen">
-    <link href="packages/codemirror/addon/lint/lint.css" rel="stylesheet" media="screen">
-    <link href="packages/codemirror/addon/hint/show-hint.css" rel="stylesheet" media="screen">
-    <link href="packages/codemirror/theme/zenburn.css" rel="stylesheet" media="screen">
-    <link href="packages/dart_pad/editing/editor_codemirror.css" rel="stylesheet" media="screen">
-    <script src="packages/codemirror/codemirror.js" defer></script>
+    <link href="../packages/codemirror/css/codemirror.css" rel="stylesheet" media="screen">
+    <link href="../packages/codemirror/addon/lint/lint.css" rel="stylesheet" media="screen">
+    <link href="../packages/codemirror/addon/hint/show-hint.css" rel="stylesheet" media="screen">
+    <link href="cm-scrollbars.css" rel="stylesheet" media="screen">
+    <link href="cm-dartpad-dark.css" rel="stylesheet" media="screen">
+    <link href="cm-dartpad-light.css" rel="stylesheet" media="screen">
 
-    <script src="scripts/embed.dart.js" defer></script>
-    <script src="scripts/ga.js" defer></script>
-    <script src="scripts/polymerpatch.js" defer></script>
-    <style>
-      horizontal-splitter.splitter {
-        margin-right: 0 !important;
-      }
+    <script src="../packages/codemirror/codemirror.js" defer></script>
+    <script src="../packages/codemirror/addon/scroll/simplescrollbars.js" defer></script>
+    <script src="../packages/split/split.min.js" defer></script>
 
-      dartpad-console {
-        margin-right: 0;
-      }
-    </style>
-  </head>
+    <script defer src="new_embed_dart.dart.js"></script>
+    <script defer src="../scripts/ga.js"></script>
+    <script defer src="../scripts/polymerpatch.js"></script>
+    <script type="text/javascript" src="dark_mode.js"></script>
 
-  <body class="fullbleed vertical layout" unresolved>
-    <div id="main" class="flex layout horizontal">
-      <paper-progress id="run-progress" class="bottom fit" hidden></paper-progress>
-      <div id="leftPanel" class="relative vertical layout">
-        <dartpad-buttons></dartpad-buttons>
-        <dartpad-editor id='topPanel' dart class="layout vertical"></dartpad-editor>
-        <horizontal-splitter class="splitter" horizontal></horizontal-splitter>
-        <dartpad-console id='bottomPanel' class="flex layout vertical"></dartpad-console>
-      </div>
-      <vertical-splitter class="splitter layout horizontal" vertical></vertical-splitter>
-      <dartpad-documentation id="rightPanel" class="flex layout vertical"></dartpad-documentation>
+    <!-- begin Material Design -->
+    <script src="../packages/mdc_web/material-components-web.min.js"></script>
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!-- end Material Design -->
+</head>
 
-      <paper-toast id="errorToast" text=""></paper-toast>
+<body class="d-flex flex-column">
+<div id="progress-bar" role="progressbar" class="mdc-linear-progress hide">
+    <div class="mdc-linear-progress__buffering-dots"></div>
+    <div class="mdc-linear-progress__buffer"></div>
+    <div class="mdc-linear-progress__bar mdc-linear-progress__primary-bar">
+        <span class="mdc-linear-progress__bar-inner"></span>
     </div>
+    <div class="mdc-linear-progress__bar mdc-linear-progress__secondary-bar">
+        <span class="mdc-linear-progress__bar-inner"></span>
+    </div>
+</div>
+<div id="navbar">
+    <nav>
+        <div class="mdc-tab-bar" role="tablist">
+            <div class="mdc-tab-scroller">
+                <div class="mdc-tab-scroller__scroll-area">
+                    <div class="mdc-tab-scroller__scroll-content">
+                        <button id="editor-tab" class="mdc-tab mdc-tab--active" role="tab" aria-selected="true"
+                                tabindex="0">
+                             <span class="mdc-tab__content">
+                                 <span class="mdc-tab__text-label">Dart</span>
+                             </span>
+                            <span class="mdc-tab-indicator mdc-tab-indicator--active">
+                                 <span class="mdc-tab-indicator__content mdc-tab-indicator__content--underline"></span>
+                             </span>
+                            <span class="mdc-tab__ripple"></span>
+                        </button>
+                        <button id="solution-tab" class="mdc-tab mdc-tab" role="tab" tabindex="0" hidden>
+                             <span class="mdc-tab__content">
+                                 <span class="mdc-tab__text-label">Solution</span>
+                             </span>
+                            <span class="mdc-tab-indicator mdc-tab-indicator">
+                                 <span class="mdc-tab-indicator__content mdc-tab-indicator__content--underline"></span>
+                             </span>
+                            <span class="mdc-tab__ripple"></span>
+                        </button>
+                        <button id="test-tab" class="mdc-tab mdc-tab" role="tab" tabindex="0">
+                             <span class="mdc-tab__content">
+                                 <span class="mdc-tab__text-label">Tests</span>
+                            </span>
+                            <span class="mdc-tab-indicator mdc-tab-indicator">
+                                 <span class="mdc-tab-indicator__content mdc-tab-indicator__content--underline"></span>
+                             </span>
+                            <span class="mdc-tab__ripple"></span>
+                        </button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </nav>
+    <div class="nav-buttons">
+        <button id="show-hint" class="mdc-button">Hint</button>
+        <button id="format-code" class="mdc-button">Format</button>
+        <button id="reload-gist" class="mdc-button">Reset</button>
+        <button id="execute" class="mdc-button mdc-button--raised">
+            <i class="material-icons mdc-button__icon">play_arrow</i>
+            Run
+        </button>
+        <div class="position-relative text-right pr-2">
+            <button id="menu-button" class="mdc-icon-button material-icons">more_vert</button>
+            <div id="main-menu" class="mdc-menu mdc-menu-surface">
+                <ul class="mdc-list" role="menu" aria-hidden="true" aria-orientation="vertical" tabindex="-1">
+                    <li id="show-test-menu-item" class="mdc-list-item" role="menuitem">
+                        <span id="show-test-checkmark" class="mdc-list-item__graphic hide">
+                            <i class="material-icons mdc-select__icon" tabindex="-1" role="button">check</i>
+                        </span>
+                        <span class="mdc-list-item__text">Show tests</span>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </div>
+</div>
+<section id="tab-container">
+    <div id="editor-container">
+        <div id="user-code-view" class="flex-auto d-flex flex-row no-overflow">
+            <div id="user-code-editor" spellcheck="false"></div>
+        </div>
+        <div id="solution-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
+            <div id="solution-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
+        </div>
+        <div id="test-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
+            <div id="test-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
+        </div>
+    </div>
+    <div id="console-view" hidden>
+        <div class="console-output-header">
+            <span class="view-label">Console <span id="unread-console-counter" class="Counter" hidden></span></span>
+        </div>
+        <div id="console-output-container" class="custom-scrollbar" hidden></div>
+    </div>
+    <div id="web-output" hidden>
+        <iframe id="frame" sandbox="allow-scripts" class="flex-auto">
+        </iframe>
+    </div>
+</section>
+<footer flex layout horizontal class="footer">
+    <div id="issues-message"></div>
+    <button id="issues-toggle" class="mdc-button mdc-button--dense mdc-button--outlined" hidden></button>
+</footer>
 
-    <dartpad-dom-output hidden class="flex"></dartpad-dom-output>
-    <reset-dialog></reset-dialog>
-    <export-dialog></export-dialog>
-  </body>
+<div id="flash-container" class="position-fixed right-2 bottom-6">
+    <div id="analysis-result-box" class="flash flash-error" hidden>
+        <button class="flash-close">
+            <div class="octicon octicon-x"></div>
+        </button>
+        <div class="message-container custom-scrollbar"></div>
+    </div>
+    <div id="test-result-box" class="flash flash-warn" hidden>
+        <button class="flash-close">
+            <div class="octicon octicon-x"></div>
+        </button>
+        <div class="message-container custom-scrollbar"></div>
+    </div>
+    <div id="hint-box" class="flash" hidden>
+        <button class="flash-close">
+            <div class="octicon octicon-x"></div>
+        </button>
+        <div class="message-container custom-scrollbar"></div>
+    </div>
+    <div id="issues" hidden></div>
+</div>
+
+<div class="mdc-dialog"
+     role="alertdialog"
+     aria-modal="true"
+     aria-labelledby="my-dialog-title"
+     aria-describedby="my-dialog-content">
+    <div class="mdc-dialog__container">
+        <div class="mdc-dialog__surface">
+            <h2 class="mdc-dialog__title" id="my-dialog-title"></h2>
+            <div class="mdc-dialog__content" id="my-dialog-content"></div>
+            <footer class="mdc-dialog__actions">
+                <button type="button" id="dialog-left-button" class="mdc-button mdc-dialog__button">
+                    <span class="mdc-button__label">No</span>
+                </button>
+                <button type="button" id="dialog-right-button" class="mdc-button mdc-dialog__button">
+                    <span class="mdc-button__label">Yes</span>
+                </button>
+            </footer>
+        </div>
+    </div>
+    <div class="mdc-dialog__scrim"></div>
+</div>
+
+</body>
+
 </html>