--- a/web/embed-flutter.html
+++ b/web/embed-flutter.html
@@ -33,7 +33,7 @@ BSD-style license that can be found in the LICENSE file. -->
     <script src="../packages/codemirror/addon/scroll/simplescrollbars.js" defer></script>
     <script src="../packages/split/split.min.js" defer></script>
 
-    <script defer src="new_embed_inline.dart.js"></script>
+    <script defer src="new_embed_flutter.dart.js"></script>
     <script defer src="../scripts/ga.js"></script>
     <script defer src="../scripts/polymerpatch.js"></script>
     <script type="text/javascript" src="dark_mode.js"></script>
@@ -117,29 +117,38 @@ BSD-style license that can be found in the LICENSE file. -->
         </div>
     </div>
 </div>
-<section id="tab-container" class="flex-column">
-    <div id="editor-container">
-        <div id="user-code-view" class="flex-auto d-flex flex-row no-overflow">
-            <div id="user-code-editor" spellcheck="false"></div>
+<section id="tab-container">
+    <div id="editor-and-console-container">
+        <div id="editor-container">
+            <div id="user-code-view" class="flex-auto d-flex flex-row no-overflow">
+                <div id="user-code-editor" spellcheck="false"></div>
+            </div>
+            <div id="solution-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
+                <div id="solution-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
+            </div>
+            <div id="test-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
+                <div id="test-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
+            </div>
         </div>
-        <div id="solution-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
-            <div id="solution-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
-        </div>
-        <div id="test-view" class="flex-auto d-flex flex-row no-overflow" spellcheck="false" hidden>
-            <div id="test-editor" class="flex-auto d-flex flex-column" spellcheck="false"></div>
+        <div id="console-output-footer" class="footer-top-border" hidden>
+            <div class="console-output-header">
+                <span class="view-label">
+                    Console <span id="unread-console-counter" class="Counter" hidden></span>
+                </span>
+                <button id="console-expand-button" class="flash-close">
+                    <div id="console-expand-icon" class="octicon octicon-triangle-up"></div>
+                </button>
+            </div>
+            <div id="console-output-container" class="custom-scrollbar" hidden></div>
         </div>
     </div>
-    <div id="console-view" hidden>
-        <div class="console-output-header">
-            <span class="view-label">Console <span id="unread-console-counter" class="Counter" hidden></span></span>
-        </div>
-        <div id="console-output-container" class="custom-scrollbar" hidden></div>
-    </div>
-    <div id="web-output" hidden>
+    <div id="web-output">
+        <span id="web-output-label" class="position-absolute view-label">UI Output</span>
         <iframe id="frame" sandbox="allow-scripts" class="flex-auto">
         </iframe>
     </div>
 </section>
+
 <footer flex layout horizontal class="footer">
     <div id="issues-message"></div>
     <button id="issues-toggle" class="mdc-button mdc-button--dense mdc-button--outlined" hidden></button>
@@ -190,4 +199,5 @@ BSD-style license that can be found in the LICENSE file. -->
 </div>
 
 </body>
+
 </html>