--- a/web/embed-html.html
+++ b/web/embed-html.html
@@ -19,27 +19,27 @@ BSD-style license that can be found in the LICENSE file. -->
     <link href='https://fonts.googleapis.com/css?family=Inconsolata:400,700'
           rel='stylesheet' type='text/css'>
 
-    <link rel="stylesheet" type="text/css" href="../packages/octicons_css/octicons_4.4.0.css">
+    <link rel="stylesheet" type="text/css" href="packages/octicons_css/octicons_4.4.0.css">
 
     <!-- codemirror -->
-    <link href="../packages/codemirror/css/codemirror.css" rel="stylesheet" media="screen">
-    <link href="../packages/codemirror/addon/lint/lint.css" rel="stylesheet" media="screen">
-    <link href="../packages/codemirror/addon/hint/show-hint.css" rel="stylesheet" media="screen">
-    <link href="cm-scrollbars.css" rel="stylesheet" media="screen">
-    <link href="cm-dartpad-dark.css" rel="stylesheet" media="screen">
-    <link href="cm-dartpad-light.css" rel="stylesheet" media="screen">
+    <link href="packages/codemirror/css/codemirror.css" rel="stylesheet" media="screen">
+    <link href="packages/codemirror/addon/lint/lint.css" rel="stylesheet" media="screen">
+    <link href="packages/codemirror/addon/hint/show-hint.css" rel="stylesheet" media="screen">
+    <link href="experimental/cm-scrollbars.css" rel="stylesheet" media="screen">
+    <link href="experimental/cm-dartpad-dark.css" rel="stylesheet" media="screen">
+    <link href="experimental/cm-dartpad-light.css" rel="stylesheet" media="screen">
 
-    <script src="../packages/codemirror/codemirror.js" defer></script>
-    <script src="../packages/codemirror/addon/scroll/simplescrollbars.js" defer></script>
-    <script src="../packages/split/split.min.js" defer></script>
+    <script src="packages/codemirror/codemirror.js" defer></script>
+    <script src="packages/codemirror/addon/scroll/simplescrollbars.js" defer></script>
+    <script src="packages/split/split.min.js" defer></script>
 
-    <script defer src="new_embed_html.dart.js"></script>
-    <script defer src="../scripts/ga.js"></script>
-    <script defer src="../scripts/polymerpatch.js"></script>
+    <script defer src="experimental/new_embed_html.dart.js"></script>
+    <script defer src="scripts/ga.js"></script>
+    <script defer src="scripts/polymerpatch.js"></script>
     <script type="text/javascript" src="dark_mode.js"></script>
 
     <!-- begin Material Design -->
-    <script src="../packages/mdc_web/material-components-web.min.js"></script>
+    <script src="packages/mdc_web/material-components-web.min.js"></script>
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <!-- end Material Design -->
 </head>