… bower.json
… lib/embed/embed.dart
… lib/polymer/base.dart
… lib/polymer/iron.dart
… lib/polymer/paper.dart
… lib/polymer/polymer.dart
… web/embed-inline.html
… web/scripts/embed.dart
… web/scripts/embed_components.html
… web/scripts/imports.html