Replaces the embed-*.html files with the experimental ones. This does not make changes to the experimental/ directory yet. 

This doesn't tackle #1185, but the websites are already using the experimental embeds so I don't see it as a blocker. Will be following up soon with a fix for that.

After this, the next step will be to update the websites and codelabs to stop using `experimental/` and start using these versions, so we can remove `experimental/`.

closes #892