@johnpryan 2019-08-13
@devoncarew @domesticmouse I think this cleans out most of the polymer files and references, but there might be more. Could you take a look to see if I'm missing anything?