@domesticmouse 2019-08-13
The bit to confirm is that all the relevant pages still load and render properly. I have a niggling memory that when I attempted to cut polymer out I accidentally broke the main page.