@RedBrogdon 2019-08-14
We also need to look through our analytics data to find out who else is using the embeds and make sure their integrations won't break. We should be able to test them right now by loading the page, altering the iframe URL to point to experiemental with the same query params, and see what happens.