@devoncarew 2019-08-14
> We also need to look through our analytics data to find out who else is using the embeds and make 

That makes sense.

You all may already know the main places the embeds are being used. From memory, the main usage was the dartlang web site. cc @kwalrath 

It would also be worth getting some page load numbers from Chrome DevTools. When rolling out the previous (polymer based) embed UI, we spent much of our time on the last week or so optimizing the page load time (by asset concatenation at the time, and by tweaking the caching settings in the appengine yaml config file).