@kwalrath 2019-08-14
I believe that @johnpryan has already taken care of all the embeds in dart.dev.