@johnpryan 2019-08-14
> It would also be worth getting some page load numbers from Chrome DevTools

That's a good point. We're seeing slower page load times in GA (around 65% slower.) Created #1205 to track.