@johnpryan 2019-08-15
@domesticmouse On my machine, index.html is still working as expected.

@RedBrogdon I've looked through the analytics data and can discuss that offline

On the performance side, I did some quick tests to see how the new embed performs. For a brand new page load (empty cache + hard reload) it loads slightly faster (down from 1100ms to 740ms from initial request to the onload event and down 990ms to 578ms to the first meaningful paint).

It's unclear why we're seeing GA show a slower page load, but it looks like analytics.js is loading about 515ms faster on the new embed, so it's possible that analytics.js starts measuring the performance after it loads, so it thinks the new embed is slower.