@RedBrogdon 2019-08-19
I think it would probably be safe to fire this thing off right now, since you've done a lot of work to ensure the old embeds continue functioning. To be sure, though, we should be able to check all these boxes:

* The new embed supports every query string option that is actually used by an existing embed (e.g. `run` doesn't count).
* Our O&O sites have already been updated to the new embed's URL.
* We've looked through any other pages sending us significant traffic (>100 hits/wk, perhaps) and manually checked that they'll continue working by changing URLs in inspector.

I believe the first two of these are done, and last I heard the conversation was about knocking together a spreadsheet to spread around the work of doing the last one.