@johnpryan 2019-08-19
@RedBrogdon done, discussed offline.